Move metadata file to internals namespace
