+++ src/anonpy/internals/__init__.py
@@ -6,3 +6,9 @@ from .log_handler import LogHandler
 from .request_handler import RequestHandler
 from .timeout import Timeout
 from .utils import _progressbar_options, ignore_warnings, join_url, read_file, str2bool, unique
+from .metadata import (
+    __package__,
+    __version__,
+    __license__,
+    __credits__
+)
